FEATURE: Use DTO for prescription Model

+++ Services/PrescriptionService.cs
@@ -1,35 +1,58 @@
+using AutoMapper;
 using ubuntu_health_api.Models;
+using ubuntu_health_api.Models.DTO;
 using ubuntu_health_api.Repositories;
 
 namespace ubuntu_health_api.Services
 {
-    public class PrescriptionService(IPrescriptionRepository prescriptionRepository) : IPrescriptionService
+  public class PrescriptionService(IPrescriptionRepository prescriptionRepository, IMapper mapper) : IPrescriptionService
+  {
+    private readonly IPrescriptionRepository _prescriptionRepository = prescriptionRepository;
+    private readonly IMapper _mapper = mapper;
+
+    public async Task<IEnumerable<PrescriptionDto>> GetAllPrescriptionsAsync(string tenantId)
     {
-        private readonly IPrescriptionRepository _prescriptionRepository = prescriptionRepository;
-
-        public async Task<IEnumerable<Prescription>> GetAllPrescriptionsAsync()
-        {
-            return await _prescriptionRepository.GetAllPrescriptionsAsync();
-        }
-
-        public async Task<Prescription> GetPrescriptionByIdAsync(int id)
-        {
-            return await _prescriptionRepository.GetPrescriptionByIdAsync(id);
-        }
-
-        public async Task AddPrescriptionAsync(Prescription prescription)
-        {
-            await _prescriptionRepository.AddPrescriptionAsync(prescription);
-        }
-
-        public async Task DeletePrescriptionAsync(int id)
-        {
-            await _prescriptionRepository.DeletePrescriptionAsync(id);
-        }
-
-        public async Task UpdatePrescriptionAsync(Prescription prescription)
-        {
-            await _prescriptionRepository.UpdatePrescriptionAsync(prescription);
-        }
+      var prescriptions = await _prescriptionRepository.GetAllPrescriptionsAsync(tenantId) ??
+      throw new KeyNotFoundException("No Prescriptions Found");
+      return _mapper.Map<IEnumerable<PrescriptionDto>>(prescriptions);
     }
+
+    public async Task<PrescriptionDto> GetPrescriptionByIdAsync(int id, string tenantId)
+    {
+      var prescription = await _prescriptionRepository.GetPrescriptionByIdAsync(id, tenantId) ??
+      throw new KeyNotFoundException("Prescription not found");
+      return _mapper.Map<PrescriptionDto>(prescription);
+    }
+
+    public async Task AddPrescriptionAsync(Prescription prescription, string tenantId)
+    {
+      prescription.TenantId = tenantId;
+      await _prescriptionRepository.AddPrescriptionAsync(prescription);
+    }
+
+    public async Task<bool> DeletePrescriptionAsync(int id, string tenantId)
+    {
+      var prescription = await _prescriptionRepository.GetPrescriptionByIdAsync(id, tenantId);
+      if (prescription == null)
+      {
+        return false;
+      }
+
+      await _prescriptionRepository.DeletePrescriptionAsync(id);
+      return true;
+    }
+
+    public async Task<bool> UpdatePrescriptionAsync(Prescription prescription, string tenantId)
+    {
+      var existingPrescription = await _prescriptionRepository.GetPrescriptionByIdAsync(prescription.PrescriptionId, tenantId);
+      if (existingPrescription == null)
+      {
+        return false;
+      }
+
+      prescription.TenantId = tenantId;
+      await _prescriptionRepository.UpdatePrescriptionAsync(prescription);
+      return true;
+    }
+  }
 }
